leetcode, problem 4, 4/12/2022

+++ Helper.cs
@@ -44,7 +44,7 @@ namespace Algorithm_in_CSharp
             return ints;
         }
 
-        public static void printBefore(int[] intArr,bool printing){
+        public static void printBefore(double[] intArr,bool printing){
             if(printing == false){
                 return;
             }
